+++ packages/storage/src/index.ts
@@ -1,5 +1,5 @@
 import type { BaseStorageAdapter } from './adapters/BaseStorageAdapter.js'
-import { createUploadSavers } from './createProcessors.js'
+import { createUploadSavers } from './createSavers.js'
 import type { ModelNames, UploadsConfig } from './prismaExtension.js'
 import { createUploadsExtension } from './prismaExtension.js'
 import type { UrlSigner } from './UrlSigner.js'
